requirements.txt
# Core dependencies
fastmcp>=2.13.0
httpx>=0.24.0
pydantic>=2.0.0
starlette>=0.27.0
uvicorn>=0.23.0
fastapi>=0.104.0

# OAuth and Authentication
authlib>=1.2.0
python-dotenv>=1.0.0
itsdangerous>=2.1.0